+++ broker/cloud_functions/filter_exgalac_trans/main.py
@@ -9,7 +9,8 @@ from google.cloud import logging
 import numpy as np
 import os
 
-from broker_utils import data_utils, gcp_utils, schema_maps
+from broker_utils import data_utils, gcp_utils
+from broker_utils.schema_maps import load_schema_map, get_key, get_value
 
 
 PROJECT_ID = os.getenv("GCP_PROJECT")
@@ -26,7 +27,8 @@ ps_topic = f"{SURVEY}-exgalac_trans_cf"
 if TESTID != "False":  # attach the testid to the names
     ps_topic = f"{ps_topic}-{TESTID}"
 
-schema_map = schema_maps.load_schema_map(SURVEY, TESTID)
+schema_map = load_schema_map(SURVEY, TESTID)
+sobjectId, ssourceId = get_key("objectId", schema_map), get_key("sourceId", schema_map)
 
 
 def run(msg: dict, context) -> None:
@@ -54,8 +56,8 @@ def run(msg: dict, context) -> None:
         base64.b64decode(msg["data"]), drop_cutouts=True, schema_map=schema_map
     )
     attrs = {
-        schema_map["objectId"]: str(alert_dict[schema_map["objectId"]]),
-        schema_map["sourceId"]: str(alert_dict[schema_map["sourceId"]]),
+        sobjectId: get_value("objectId", alert_dict, schema_map),
+        ssourceId: get_value("sourceId", alert_dict, schema_map),
     }
 
     if _is_extragalactic_transient(alert_dict):