+++ docs/source/working-notes/troyraen/AllWISE/cloud_run/allwise.py
@@ -12,7 +12,8 @@ from google.cloud import logging, bigquery
 import os
 from typing import Optional
 
-from broker_utils import data_utils, gcp_utils, schema_maps
+from broker_utils import data_utils, gcp_utils
+from broker_utils.schema_maps import load_schema_map, get_key, get_value
 
 
 app = Flask(__name__)
@@ -34,7 +35,8 @@ if TESTID != "False":  # attach the testid to the names
     ps_topic = f"{ps_topic}-{TESTID}"
 bq_table = f"{bq_dataset}.xmatch"
 
-schema_map = schema_maps.load_schema_map(SURVEY, TESTID)
+schema_map = load_schema_map(SURVEY, TESTID)
+sobjectId, ssourceId = get_key("objectId", schema_map), get_key("sourceId", schema_map)
 
 
 @app.route("/", methods=["POST"])
@@ -59,8 +61,8 @@ def index():
         base64.b64decode(msg["data"]), drop_cutouts=True, schema_map=schema_map
     )
     attrs = {
-        schema_map["objectId"]: str(alert_dict[schema_map["objectId"]]),
-        schema_map["sourceId"]: str(alert_dict[schema_map["sourceId"]]),
+        sobjectId: get_value("objectId", alert_dict, schema_map),
+        ssourceId: get_value("sourceId", alert_dict, schema_map),
     }
 
     # do the cross match